requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
scikit-fuzzy
numpy
psycopg2-binary